chore: set invoice list page as home page
