+++ src/pages/index.tsx
@@ -1,32 +1,45 @@
-import Head from 'next/head';
+import { useQuery } from '@tanstack/react-query';
+import axios from 'axios';
 import { useRouter } from 'next/router';
+import QueryString from 'qs';
 import { FC } from 'react';
 
-import { useAuth } from '@src/contexts/AuthContext';
+import Typography from '@components/atoms/Typography';
+import Drawer from '@components/molecules/Drawer';
+import InvoiceListActionBar from '@components/molecules/InvoiceListActionBar';
+import CreateInvoiceForm from '@components/organisms/CreateInvoiceForm';
+import InvoiceCardList from '@components/organisms/InvoiceCardList';
+import { ETypographyVariant } from '@enums/typography';
 import CoreLayout from '@src/layouts/core';
+import { Invoice } from '@types';
 
-const Home: FC = () => {
-  const { logout, user } = useAuth();
-  const router = useRouter();
+const DEFAULT_QUERY_KEY = 'All';
+
+const InvoicesPage: FC = () => {
+  const { query } = useRouter();
+  const { data, error } = useQuery(['invoices', query.status || DEFAULT_QUERY_KEY], () =>
+    axios
+      .get<{ invoices: Invoice.IModel[] }>(`/api/invoices?${QueryString.stringify(query)}`)
+      .then((data) => data.data)
+  );
+
+  if (error) {
+    return <Typography displayAs={ETypographyVariant.H3}>Oopsy</Typography>;
+  }
+
   return (
     <CoreLayout>
-      <Head>
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <title>Billy.io</title>
-      </Head>
-      <pre>{JSON.stringify(user, undefined, 4)}</pre>
-      <button
-        onClick={() => {
-          logout();
-          router.push('/login');
-        }}
-      >
-        logout
-      </button>
-      <p>Welcome to Billy.io</p>
-      <p>This is a very empty homepage...</p>
+      <InvoiceListActionBar invoiceCount={data?.invoices.length ?? 0} />
+      {data ? (
+        <InvoiceCardList invoices={data.invoices} />
+      ) : (
+        <Typography displayAs={ETypographyVariant.H3}>Loading...</Typography>
+      )}
+      <Drawer>
+        <CreateInvoiceForm />
+      </Drawer>
     </CoreLayout>
   );
 };
 
-export default Home;
+export default InvoicesPage;
